+++ company/widgets.py
@@ -1,6 +1,6 @@
 from django.core.validators import EMPTY_VALUES
 from django.forms import widgets
-from django.utils.html import format_html, mark_safe
+from django.utils.html import format_html
 
 
 class PreventRenderWidget(widgets.Input):
@@ -9,26 +9,6 @@ class PreventRenderWidget(widgets.Input):
 
     def render(self, name, value, attrs=None, renderer=None):
         return format_html('<!- not rendered ->')
-
-
-class CheckboxWithInlineLabel(widgets.CheckboxInput):
-    template = """
-        <div class="form-field checkbox">
-            {input_html}
-            <label for="{id}">{label}</label>
-        </div>
-    """
-
-    def __init__(self, label='', *args, **kwargs):
-        self.label = label
-        super().__init__(*args, **kwargs)
-
-    def render(self, name, value, attrs=None):
-        input_html = super().render(name, value, attrs)
-        wrapper_html = self.template.format(
-            input_html=input_html, label=self.label, id=attrs['id']
-        )
-        return mark_safe(wrapper_html)
 
 
 class CheckboxSelectInlineLabelMultiple(widgets.CheckboxSelectMultiple):
