+++ core/helpers.py
@@ -1,7 +1,8 @@
+from directory_cms_client import DirectoryCMSClient
+
 from django.conf import settings
 from django.shortcuts import Http404
-
-from directory_cms_client import DirectoryCMSClient
+from django.utils import translation
 
 
 cms_client = DirectoryCMSClient(
@@ -15,3 +16,10 @@ def handle_cms_response(response):
         raise Http404()
     response.raise_for_status()
     return response.json()
+
+
+def get_language_from_querystring(request):
+    language_code = request.GET.get('lang')
+    language_codes = translation.trans_real.get_languages()
+    if language_code and language_code in language_codes:
+        return language_code
