+++ core/tests/test_context_processors.py
@@ -1,13 +1,12 @@
 from django.core.urlresolvers import reverse
 
-from ui import context_processors
-from core.forms import AnonymousSubscribeForm, LeadGenerationForm
+from core import context_processors, forms
 
 
 def test_feature_flags_installed(settings):
     processors = settings.TEMPLATES[0]['OPTIONS']['context_processors']
 
-    assert 'ui.context_processors.feature_flags' in processors
+    assert 'core.context_processors.feature_flags' in processors
 
 
 def test_feature_returns_expected_features(rf, settings):
@@ -22,7 +21,7 @@ def test_feature_returns_expected_features(rf, settings):
 def test_subscribe_form_installed(settings):
     processors = settings.TEMPLATES[0]['OPTIONS']['context_processors']
 
-    assert 'ui.context_processors.subscribe_form' in processors
+    assert 'core.context_processors.subscribe_form' in processors
 
 
 def test_subscribe_form_exposes_form_details(rf):
@@ -30,7 +29,9 @@ def test_subscribe_form_exposes_form_details(rf):
 
     actual = context_processors.subscribe_form(request)
 
-    assert isinstance(actual['subscribe']['form'], AnonymousSubscribeForm)
+    assert isinstance(
+        actual['subscribe']['form'], forms.AnonymousSubscribeForm
+    )
 
 
 def test_analytics(rf, settings):
@@ -58,7 +59,7 @@ def test_analytics_installed(settings):
 def test_lead_generation_form_installed(settings):
     processors = settings.TEMPLATES[0]['OPTIONS']['context_processors']
 
-    assert 'ui.context_processors.lead_generation_form' in processors
+    assert 'core.context_processors.lead_generation_form' in processors
 
 
 def test_lead_generation_form_exposes_form_details(rf):
@@ -66,4 +67,6 @@ def test_lead_generation_form_exposes_form_details(rf):
 
     actual = context_processors.lead_generation_form(request)
 
-    assert isinstance(actual['lead_generation']['form'], LeadGenerationForm)
+    assert isinstance(
+        actual['lead_generation']['form'], forms.LeadGenerationForm
+    )
