+++ core/tests/test_views.py
@@ -1,37 +1,8 @@
 from unittest.mock import patch
 
-import pytest
-
 from django.core.urlresolvers import reverse
 
 from core.tests import helpers
-
-
-@pytest.fixture
-def buyer_form_data():
-    return {
-        'full_name': 'Jim Example',
-        'email_address': 'jim@example.com',
-        'sector': 'AEROSPACE',
-        'company_name': 'Deutsche Bank',
-        'country': 'Germany',
-        'terms': True,
-        'comment': 'This website should be all in German.',
-    }
-
-
-@pytest.fixture
-def buyer_request(rf, client, buyer_form_data):
-    request = rf.post('/', buyer_form_data)
-    request.session = client.session
-    return request
-
-
-@pytest.fixture
-def buyer_request_invalid(rf, client):
-    request = rf.post('/', {})
-    request.session = client.session
-    return request
 
 
 @patch('core.helpers.cms_client.lookup_by_slug')
