+++ requirements.txt
@@ -13,8 +13,8 @@ chardet==3.0.4            # via requests
 contextlib2==0.5.5        # via raven
 directory-api-client==9.16.0
 directory-client-core==0.2.1  # via directory-api-client, directory-cms-client
-directory-cms-client==1.0.0
-directory_components==1.9.0
+directory-cms-client==2.0.0
+directory_components==2.1.0
 directory_constants==6.7.0
 directory_validators==4.5.0
 django-formtools==2.0
