--- ui/tests/test_sitemaps.py
@@ -1,11 +1,0 @@
-import http
-
-from django.core.urlresolvers import reverse
-
-
-def test_sitemaps_200(client):
-    url = reverse('sitemap')
-
-    response = client.get(url)
-
-    assert response.status_code == http.client.OK
